+++ src/gl_backend.h
@@ -149,8 +149,8 @@ class GLBackend
     void NewFrame();
     renderStats_t *RenderStats();
 
-    static void BindTexture(size_t unit, const GLTexture *texture);
-    static void BindTexture(size_t unit, GLuint texture);
+    static void BindTexture(int unit, const GLTexture *texture);
+    static void BindTexture(int unit, GLuint texture);
     static void SetUniformValue(ShaderUniform *it);
